refactor: 공용 컴포넌트를 atoms 폴더로 이동
