+++ src/components/atoms/text.ts
